# api_gateway/api/docs/swagger.yaml
definitions:
...
  models.ResponseModel:
    properties:
      code:
        type: integer
      data: {}
      error: {}
      message:
        type: string
    type: object
  models.Status:
    properties:
      status:
        type: string
    type: object
...
  user_service.Branch:
    properties:
      active:
        type: boolean
      address:
        type: string
      created_at:
        type: string
      delivery_tarif_id:
        type: integer
      destination:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      photo:
        type: string
      updated_at:
        type: string
      work_hour_end:
        type: string
      work_hour_start:
        type: string
    type: object
...
  user_service.CreateBranchRequest:
    properties:
      address:
        type: string
      delivery_tarif_id:
        type: integer
      destination:
        type: string
      name:
        type: string
      phone:
        type: string
      photo:
        type: string
      work_hour_end:
        type: string
      work_hour_start:
        type: string
    type: object
...
  user_service.ListBranchResponse:
    properties:
      branches:
        items:
          $ref: '#/definitions/user_service.Branch'
        type: array
      count:
        type: integer
    type: object
...
  user_service.Response:
    properties:
      message:
        type: string
    type: object
  user_service.UpdateBranchRequest:
    properties:
      address:
        type: string
      delivery_tarif_id:
        type: integer
      destination:
        type: string
      id:
        type: integer
      name:
        type: string
      phone:
        type: string
      photo:
        type: string
      work_hour_end:
        type: string
      work_hour_start:
        type: string
    type: object
...
info:
  contact: {}
paths:
  /v1/branch/create:
    post:
      consumes:
      - application/json
      description: create branch
      operationId: create-branch
      parameters:
      - description: branch
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/user_service.CreateBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/user_service.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: create branch
      tags:
      - branch
  /v1/branch/delete/{branch_id}:
    delete:
      consumes:
      - application/json
      description: Delete Branch
      operationId: delete-branch
      parameters:
      - description: branch_id
        in: path
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.Status'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: delete branch
      tags:
      - branch
  /v1/branch/get/{branch_id}:
    get:
      consumes:
      - application/json
      description: get branch
      operationId: get-branch-byID
      parameters:
      - description: branch_id
        in: path
        name: branch_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/user_service.Branch'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get branch by ID
      tags:
      - branch
  /v1/branch/list:
    get:
      consumes:
      - application/json
      description: get branch
      operationId: get-branch
      parameters:
      - description: search
        in: query
        name: search
        type: string
      - description: limit
        in: query
        name: limit
        type: integer
      - description: page
        in: query
        name: page
        type: integer
      - description: name
        in: query
        name: name
        type: integer
      - description: created_from
        in: query
        name: created_from
        type: string
      - description: created_to
        in: query
        name: created_to
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: desc
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/user_service.ListBranchResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: get list branch
      tags:
      - branch
  /v1/branch/update/{branch_id}:
    put:
      consumes:
      - application/json
      description: Update Branch by ID
      operationId: update_branch
      parameters:
      - description: branch
        in: body
        name: branch
        required: true
        schema:
          $ref: '#/definitions/user_service.UpdateBranchRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                data:
                  $ref: '#/definitions/models.Status'
              type: object
        "400":
          description: Bad Request
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
        "500":
          description: Server Error
          schema:
            allOf:
            - $ref: '#/definitions/models.ResponseModel'
            - properties:
                error:
                  type: string
              type: object
      summary: Update Branch
      tags:
      - branch
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"